+++ packages/notices/src/store/actions.js
@@ -46,6 +46,11 @@ export function* createNotice( statusOrNotice = DEFAULT_STATUS, content, options
 		actions = [],
 	} = options;
 
+	// The supported value shape of content is currently limited to plain text
+	// strings. To avoid setting expectation that e.g. a WPElement could be
+	// supported, cast to a string.
+	content = String( content );
+
 	yield { type: 'SPEAK', message: content };
 
 	yield {