+++ packages/notices/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - The `createNotice` can now optionally accept a WPNotice object as the sole argument.
 
+### Bug Fixes
+
+- While `createNotice` only explicitly supported content of type `string`, it was not previously enforced. This has been corrected.
+
 ## 1.0.5 (2018-11-15)
 
 ## 1.0.4 (2018-11-09)